add component for history input
